refactor: components, styles 폴더로 파일 정리
