+++ src/pages/message/styles/editor.css
@@ -1,7 +1,7 @@
 @import url('https://fonts.googleapis.com/css2?family=Noto+Sans:wght@300;400;700&display=swap');
 @import url('https://fonts.googleapis.com/earlyaccess/nanummyeongjo.css');
 @import url('https://cdn.jsdelivr.net/gh/fonts-archive/NanumSonPyeonJiCe/NanumSonPyeonJiCe.css');
-@import '../../index.css';
+@import '../../../index.css';
 
 .ql-toolbar.ql-snow {
   @apply bg-[#eeeeee] rounded-t-lg;
